+++ test/single_model_file/models/mod.rs
@@ -1,0 +1,2 @@
+pub mod table1;
+pub mod table2;