+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - replace dependency `Inflector` with `heck`
 - add option `create-str` to set `Create*` structs string type
 - add option `update-str` to set `Update*` structs string type
+- add option `--single-model-file` to only generate a single file instead of a directory with `mod.rs` and `generated.rs`
 
 ## 0.0.16